# perf: use a buffer pool to reduce allocations