licenses size is known so we can make a pool to reuse the byte slice.
We only read at most len(license) bytes when checking the file header.

Benchmarks:

```
name        old time/op    new time/op    delta
Run/dot-16     443µs ±11%     431µs ±14%     ~     (p=0.549 n=9+10)

name        old alloc/op   new alloc/op   delta
Run/dot-16    78.0kB ± 0%    32.8kB ± 0%  -58.00%  (p=0.000 n=10+10)

name        old allocs/op  new allocs/op  delta
Run/dot-16       729 ± 0%       714 ± 0%   -2.06%  (p=0.000 n=10+10)
```